feat: add support for FSDP
Aldo Pareja implemented support for FSDP in the training library which currently
lives on the branch ap/fsdp-port. This commit ports over that code and provides
a way for consumers to now select which distributed training backend they want
to train with. This can be specified by a new field on the `TrainingArguments
object: `TrainingArguments.distributed_training_backend` which accepts a value
of "fsdp" or "deepspeed", and then runs the appropriate training loop.

This commit also sections up the training code into two separate files
so that the dependencies of deepspeed are isolated from FSDP, allowing users
who only want to use FSDP to avoid needing to pull in unnecessary dependencies.

Signed-off-by: Oleg S <97077423+RobotSail@users.noreply.github.com>